+++ local-development/devtool/adapters/registry_loader.py
@@ -15,6 +15,14 @@ except ImportError:
     )
     raise SystemExit(1)
 
+from devtool.domain.models import (
+    InfraComponent,
+    Service,
+    ServiceGroup,
+    ServiceType,
+    VenvConfig,
+    VenvStrategy,
+)
 from devtool.domain.registry import Registry
 
 _DEFAULT_YAML = Path(__file__).resolve().parent.parent.parent / "services.yaml"
@@ -37,18 +45,83 @@ class YamlRegistryLoader:
 
         min_override = (os.environ.get("PYTHON_MIN", "").strip() or None)
         max_override = (os.environ.get("PYTHON_MAX", "").strip() or None)
-        python_min, python_max = Registry.parse_python_bounds(
+        python_min, python_max = YamlRegistryLoader._parse_python_bounds(
             raw, min_override=min_override, max_override=max_override,
         )
 
         return Registry(
-            services=Registry.parse_services(raw.get("services", {})),
-            infra=Registry.parse_infra(raw.get("infrastructure", {})),
-            groups=Registry.parse_groups(raw.get("groups", {})),
+            services=YamlRegistryLoader._parse_services(raw.get("services", {})),
+            infra=YamlRegistryLoader._parse_infra(raw.get("infrastructure", {})),
+            groups=YamlRegistryLoader._parse_groups(raw.get("groups", {})),
             local_auth=local_auth,
             python_min=python_min,
             python_max=python_max,
             log_dir=Path(
                 raw.get("logging", {}).get("directory", "/tmp/unifai-dev/logs")
             ),
+        )
+
+    # -- parsing helpers (raw dict → domain model transforms) ----------------
+
+    @staticmethod
+    def _parse_infra(raw: dict) -> dict[str, InfraComponent]:
+        result: dict[str, InfraComponent] = {}
+        for name, data in raw.items():
+            result[name] = InfraComponent(
+                name=name,
+                image=data["image"],
+                ports=data.get("ports", []),
+                label=data.get("label", name),
+                command=data.get("command"),
+                stop_timeout=data.get("stop_timeout"),
+            )
+        return result
+
+    @staticmethod
+    def _parse_services(raw: dict) -> dict[str, Service]:
+        result: dict[str, Service] = {}
+        for name, data in raw.items():
+            venv_raw = data.get("venv", {})
+            venv = VenvConfig(
+                strategy=VenvStrategy(venv_raw.get("strategy", "none")),
+                commands=venv_raw.get("commands", []),
+            )
+            result[name] = Service(
+                name=name,
+                directory=Path(data["directory"]),
+                port=data.get("port"),
+                host=data.get("host"),
+                health_endpoint=data.get("health_endpoint"),
+                type=ServiceType(data.get("type", "python")),
+                infrastructure=data.get("infrastructure", []),
+                is_primary=data.get("is_primary", True),
+                env_file=data.get("env_file"),
+                env_entries=data.get("env_entries", {}),
+                venv=venv,
+                launch=data["launch"],
+            )
+        return result
+
+    @staticmethod
+    def _parse_groups(raw: dict) -> dict[str, ServiceGroup]:
+        return {
+            name: ServiceGroup(name=name, services=svc_list)
+            for name, svc_list in raw.items()
+        }
+
+    @staticmethod
+    def _parse_python_bounds(
+        raw: dict,
+        *,
+        min_override: str | None = None,
+        max_override: str | None = None,
+    ) -> tuple[tuple[int, int], tuple[int, int]]:
+        py = raw.get("python", {})
+        min_str = min_override or py.get("min", "3.11")
+        max_str = max_override or py.get("max", "3.13")
+        min_parts = min_str.split(".")
+        max_parts = max_str.split(".")
+        return (
+            (int(min_parts[0]), int(min_parts[1])),
+            (int(max_parts[0]), int(max_parts[1])),
         )
